rename outerbase => ezql
